finished hw 30.2 - Serializers
