+++ config/urls.py
@@ -4,4 +4,5 @@ from django.urls import include, path
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("lms.urls", namespace="lms")),
+    path("users/", include("users.urls", namespace="users")),
 ]
