+++ lms/serializers.py
@@ -2,17 +2,6 @@ from rest_framework.fields import SerializerMethodField
 from rest_framework.serializers import ModelSerializer
 
 from lms.models import Course, Lesson
-from users.models import CustomUser, Payment
-
-
-class UserSerializer(ModelSerializer):
-    """
-    Сериализатор для модели пользователя.
-    """
-
-    class Meta:
-        model = CustomUser
-        fields = ("id", "email")
 
 
 class LessonSerializer(ModelSerializer):
@@ -43,17 +32,3 @@ class CourseSerializer(ModelSerializer):
         Дополнительное поле, вычисляет количество уроков на курсе.
         """
         return course.lessons.count()
-
-
-class PaymentSerializer(ModelSerializer):
-    """
-    Сериализатор для модели платежей.
-    """
-    client = UserSerializer(read_only=True)
-
-    course_paid = CourseSerializer()
-    lesson_paid = LessonSerializer()
-
-    class Meta:
-        model = Payment
-        fields = "__all__"
