+++ users/serializers.py
@@ -1,7 +1,6 @@
-from rest_framework.fields import SerializerMethodField
 from rest_framework.serializers import ModelSerializer
 
-from lms.models import Course, Lesson
+from lms.serializers import CourseSerializer, LessonSerializer
 from users.models import CustomUser, Payment
 
 
@@ -13,36 +12,6 @@ class UserSerializer(ModelSerializer):
     class Meta:
         model = CustomUser
         fields = ("id", "email")
-
-
-class LessonSerializer(ModelSerializer):
-    """
-    Сериализатор для модели урока.
-    """
-
-    class Meta:
-        model = Lesson
-        fields = "__all__"
-
-
-class CourseSerializer(ModelSerializer):
-    """
-    Сериализатор для модели курса.
-    """
-
-    lessons_count = SerializerMethodField()
-    lessons = LessonSerializer(many=True)
-
-    class Meta:
-        model = Course
-        fields = "__all__"
-
-    @staticmethod
-    def get_lessons_count(course):
-        """
-        Дополнительное поле, вычисляет количество уроков на курсе.
-        """
-        return course.lessons.count()
 
 
 class PaymentSerializer(ModelSerializer):
